exams: rework finder tool, compact landing layout
