+++ apps/web/src/components/content/link.tsx
@@ -1,7 +1,7 @@
 import { faUpRightFromSquare } from '@fortawesome/free-solid-svg-icons'
 import { default as NextLink } from 'next/link'
 import { useRouter } from 'next/router'
-import { ForwardedRef, forwardRef, ReactNode } from 'react'
+import { ForwardedRef, forwardRef, ReactNode, MouseEvent } from 'react'
 
 import { FaIcon } from '../fa-icon'
 import { cn } from '@/helper/cn'
@@ -14,7 +14,7 @@ export interface LinkProps {
   forceNoCSR?: boolean //
   unreviewed?: boolean // e.g. user profiles or comments
   title?: string
-  onClick?: () => void
+  onClick?: (e: MouseEvent) => void
 }
 
 // imported from cf worker redirects.ts
